Add support for channel restrictions
Implement a whitelist of channels to collect links from in order to
limit the amount of unwanted links like bugzilla or forums on FLOSS
community channels.

Maybe a filter based on parts of the URL or simply leaving this up to
the collectr server would be a better idea but this is fine for me now.